Card dealing, Improved menu, beginning of card dragging
Started functionality of the columns and how they handle cards and stacks when they are grabbed.
Added "New Game" feature, but it only randomly places the cards right now.
Also added "Resume" button to menu, and non-functional "View Stats" button.
Also, the menu background is now perfectly sized for 5 buttons.
Clicking and dragging is very broken, however it has been started. There is a half-baked attempt to use a singleton from online that makes it possible to pick only the topmost visible card, however it does not work at the moment.