Implemented card dragging
With the help of the modified singleton we found, stacks of cards can now be dragged. No logic has been implemented yet, so you cannot place any cards anywhere yet.
As for technical details, I had to change the code to be based around the idea of stacks of cards in order to position them while dragging properly.
Also moved the card spacing number into a variable in "Main/Table" for use in multiple places.
Also fixed a bug where the game would crash if you tried clicking on a card that was currently being dragged. (E.g. dragging with mouse, but then clicking via mousepad at the same time.)