Dragging logic: stacks by color and descending number
Added outline to cards for better clarity.
Cards can only be dragged in a stack if they are alternating in color and descending in number. Cards can only be placed on top of a card that is greater than them by one value.
Also changed column's add_card function to be data type relative for ease of use in the place_stack function later.