Implemented Undo and Replay features
Undo can be used by doing ctrl+z
Replay can be used by doing ctrl+r
The make_move function in card is even more generic since it now requires a destination to be specified.
Some code was introduced to test potential move animations, but it was temporarily removed since it not fully functional.
Hands are now generated by shuffling a list of numbers from 0-51 instead of directly shuffling the cards array. This should make it easier to add numbered deals later.
The foundation now properly returns a card when a card is removed from it. Undo forces it to behave as if it's still possible to remove a card.