Numbered deals, Updated menu options
Numbered deals can be chosen from the menu. The "New Game" button has a text box on it that can accept a number. A random deal will be chosen if no seed is specified. While the text box is in focus, S and F inputs are blocked.
Added Replay option to the menu.
Settings menu now blocks game inputs like F2 and Ctrl+Z, Ctrl+Y, or Ctrl+R.
Fixed memory issue where the current_hand array was not being cleared when a new game is started.
Set up structure for detecting wins.
Rearranged the functions in settings_menu to match the menu order.
Replaced background of settings menu to be a color box instead of an imported asset. This should make my life easier when I want to change it.