Touch Screen Campatability
Buttons have been added to the screen for Undo, Redo, and Replay.
Touch now works as expected. Tapping a card will simply right-click the card. Dragging however works the same as with the mouse.
Disabled context menu on the Numbered Deal text box since it is unnecessary, however copying and highlighting is still possible.