Added Info screen, several optimizations
A new info screen displays shortcuts for the game, as well as a list of stats (stats not yet implemented). It also has a link to the GitHub repo.
Also added a button to the numbered deal text box that will generate a random number and place it in the text box.
Added focusing abilities to the menu that can be used with the tab key or arrow keys.
Cards do not shake when they are in the foundation anymore.
Improved the terrible system of detecting clicks while UI screens are open. I switched to control nodes which are capable of blocking mouse clicks when I need them to.
Also changed the settings menu to be a Control node instead of a Node2D. This had no immediate effect, but it seemed to make more sense.
Improvement on how the time is paused when screens are open. It should be more reliable.
Fixed the font of the timer, deal indicator, and move counter to match the rest of the game.