Added persistent save data, created game icon
The should now load up the player's stats on startup. if there is no data file or if the datafile is broken, a new one will be created.
Created an icon for the game. The project settings have been updated so that the icon shows on the taskbar and on the top left of the window (for windows at least).
A game will only be counted as a loss when the player has made moves on a hand and then starts a new game. Even if a hand with progress is restarted and then the player chooses to start a new game, a loss will be counted.
Added structure for confirmation screens.
Updated README to include image.