Added confirmation screens, added auto-complete button
Added confirmation for these events:
Starting a new game
Autocomplete
Resetting stats
When the game becomes autocomplete-able, a button will show up on the table. If the player decides to reject the autocomplete prompt, they can always trigger it again by clicking on the auto complete button.
Fixed a few issues with how screens are displayed and added many checks to make sure that things can't happen when they're not supposed to. When a prompt is on the screen, almost nothing should be possible to click on.
Cards should not be able to be touched or influenced while autocompleting.