Finished custom game screen
Added the buttons necessary to select cards visually from the custom game screen.
The custom game screen will keep the display in sync whether cards are changed through text or through the card selector.
Optimized some of the ordering of how the custom game screen is validated.
Moved the button to activate the custom game screen into the DarkOverlay object in the GUI to make sure it is only clickable when my name is visible.
Moved custom_card_area to the UI folder.
Rearranged function in table.gd in order of importance to make it a little easier to follow.
Added many comments.