Improved custom screen focusing again
Read the change to the README for clarification.
The arrow keys can be used to change the focus around anywhere. Using "Tab" however will alternate between moving the card location selector, and card selection panel. Using the mouse to select a card from the card panel will keep the focus on the card location selector so you can do a Click-ArrowLeft-Click-ArrowLeft type of action repeatedly.