Integrated Microsoft FreeCell random number generator
Big help from https://rosettacode.org/wiki/Deal_cards_for_FreeCell
Numbered deals from 1-9999999 are now compatible with most published FreeCell games. Deal 10,000,000 is not included, but this is simply because the digit count on the settings menu is limited to 7. Deal number 0 is just random.
Also removed the old system of seeding since it was replaced entirely. Hopefully this doesn't break things.