Fixed issue with game label
When replaying a game, the label at the top of the screen would clear out in situations that were not necessary. If you replay a game after running a simulation, it will be marked as a custom game. So right now, replaying a win blocks users from being able to replay the game with statistics, UNLESS it's a numbered deal and they remember the number.