Reduced CPU usage
The main settings that helped was the "Low Processor Mode" setting. When this settings is enabled, the screen will not refresh if nothing is happening. In most solitaire games, nothing is happening visually most of the time, so this is beneficial to me.
The other settings I changed were mostly experimental, however they reportedly reduce RAM usage, so I'll keep them there unless something starts breaking.